Fixes 2 bugs and adds new search feature.
Fixes 1 bugs in CoffeeSite creation and 1 bug in CoffeeSites
search process.
Adds new feature for searching CoffeeSites based on city/town name.
Now, if CoffeeSites in city/town are requested, first the api.mapy.cz
is asked for coordinates of this city/town and the list of found
cities is proposed to user to choose from. Then the CoffeeSites
are found based on coordinates of chosen city.